python_script/requirements.txt
pyserial
colorama
limepy
pyft4222
matplotlib
